Refactored to MVC Structure
Created Models, Controllers, and Services packages to follow MVC class structure
