+++ src/main/java/surveyMonkey/models/Responses.java
@@ -1,4 +1,4 @@
-package surveyMonkey;
+package surveyMonkey.models;
 
 import lombok.Getter;
 import lombok.Setter;
