+++ src/main/java/surveyMonkey/services/DatabaseService.java
@@ -1,0 +1,8 @@
+package surveyMonkey.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class DatabaseService {
+
+}
